pyproject.toml
[project]
name = "ocrroo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.1",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyqt6>=6.10.0",
    "pyqt6-webengine>=6.10.0",
    "pytesseract>=0.3.13",
    "python-vlc>=3.0.21203",
]
